都庁来庁者数の推移: Rename components/AgencyBarChart.vue to components/index/CardsReference/Agency/Chart.vue
